+++ changelog.md
@@ -1,9 +1,13 @@
 # Librarian changelog
 
+## 0.2.0
+
+- Add test method to evaluable.
+
 ## 0.1.0
 
 - Add probe implementation, on evaluation event.
-- Add team into the evaluation result object
+- Add team into the evaluation result object.
 
 ## 0.0.1